Move utitlity function where it is needed only
