+++ packages/utils/src/utils-index.ts
@@ -12,7 +12,6 @@ export { default as pipePromises } from './pipe-promises';
 export { default as generateRandomHash } from './generate-random-hash';
 export { default as randomValueHex } from './random-value-hex';
 export { default as buildSortOptions } from './db/build-sort-option';
-export { default as convertTagsToLowerCase } from './covert-tags-to-lower-case';
 
 /*
  * Db utils
